requirements.txt
Pillow==8.2.0
numpy==1.20.3
pandas==1.2.4
scipy==1.6.3
tqdm==4.60.0